requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# added requirements
azure-cosmos==3.0.2
beautifulsoup4
jupyterlab
pandas
requests
spacy==2.0.18
fuzzywuzzy==0.17.0
python-levenshtein==0.12.0
gremlinpython==3.2.7
hug
uwsgi
